+++ rust/src/consumers/stats.rs
@@ -20,7 +20,7 @@ pub struct GateStats {
     pub mul_gates: usize,
     pub add_constant_gates: usize,
     pub mul_constant_gates: usize,
-    pub variables_freed: u64,
+    pub variables_deleted: u64,
 
     pub functions_defined: usize,
     pub functions_called: usize,
@@ -157,9 +157,9 @@ impl GateStats {
                 self.private_variables += 1;
             }
 
-            Free(_field_id, first, last) => {
+            Delete(_field_id, first, last) => {
                 let last_one = last.unwrap_or(*first);
-                self.variables_freed += last_one - *first + 1;
+                self.variables_deleted += last_one - *first + 1;
             }
 
             Convert(_, _) => {
@@ -194,7 +194,7 @@ impl GateStats {
         self.mul_gates += other.mul_gates;
         self.add_constant_gates += other.add_constant_gates;
         self.mul_constant_gates += other.mul_constant_gates;
-        self.variables_freed += other.variables_freed;
+        self.variables_deleted += other.variables_deleted;
 
         self.convert_gates += other.convert_gates;
         self.functions_called += other.functions_called;
@@ -226,7 +226,7 @@ fn test_stats() -> Result<()> {
             mul_gates: 4,
             add_constant_gates: 0,
             mul_constant_gates: 0,
-            variables_freed: 12,
+            variables_deleted: 12,
             functions_defined: 1,
             functions_called: 3,
             convert_gates: 0,