+++ Changelog.md
@@ -12,6 +12,7 @@ Rust:
 - remove Switch
 - remove For loop
 - rename Instance/Witness by PublicInputs/PrivateInputs
+- rename Free gate by Delete gate
 
 # SIEVE IR v1.0.1, 2022-08